feat(front) adicionado novas variaveis de cor e movido sidebar
adicionado status online e offline
