+++ vue/src/layouts/shadcn/index.vue
@@ -10,350 +10,7 @@
                     <PanelLeftClose />
                 </button>
             </div>
-            <div class="flex-1 overflow-y-auto bg-sidebar text-white px-2 py-4">
-                <nav class="grid gap-2">
-                    <router-link to="/dashboard"
-                        class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                        <House />
-                        <span>Dashboard</span>
-                    </router-link>
-                    <collapsible v-if="menuStore.permissions.admin" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <LockKeyhole />
-                            <span>Administração</span>
-                            <ChevronRight class="ml-auto h-4 w-4 transition-transform" :class="{ 'rotate-90': open }" />
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/administracao/usuarios"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <Users />
-                                    <span>Usuários</span>
-                                </router-link>
-                                <router-link to="/assinatura"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <Crown />
-                                    <span>Assinatura</span>
-                                    <div class="ml-auto flex space-x-1 items-center">
-                                        <Badge :class="loginStore.isProAccount ? 'bg-orange-400 text-black' : 'bg-gray-500'"> {{
-                                            loginStore.isProAccount ? 'Pro' : 'Free' }} </Badge>
-                                    </div>
-                                </router-link>
-                                <router-link to="/administracao/logs"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <List />
-                                    <span>Logs</span>
-                                </router-link>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                    <collapsible v-if="menuStore.permissions.assistente" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <Cpu />
-                            <span>Assistente de IA</span>
-                            <div class="ml-auto flex space-x-1 items-center">
-                                <Skeleton v-if="true" class="bg-blue-500 rounded-full h-3 w-3 shadow-none" />
-                                <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                    :class="{ 'rotate-90': open }" />
-                            </div>
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/agentesia/playground"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <MessageCircle />
-                                    <span>Chat Integrado ✨</span>
-                                </router-link>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                    <collapsible v-if="menuStore.permissions.crm" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <Layers />
-                            <span>CRM</span>
-                            <div class="ml-auto flex space-x-1 items-center">
-                                <Badge v-if="true" variant="destructive"> {{ stateDevelopment }} </Badge>
-                                <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                    :class="{ 'rotate-90': open }" />
-                            </div>
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/patrimonio/categorias"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileChartPie />
-                                    <span>Resumo</span>
-                                </router-link>
-                                <router-link to="/crm/clientes"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <User />
-                                    <span>Clientes</span>
-                                </router-link>
-                                <router-link to="/vendas/pdv"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <ClipboardCheck />
-                                    <span>Relatórios</span>
-                                </router-link>
-                                <collapsible v-slot="{ open }" class="grid gap-2">
-                                    <collapsible-trigger
-                                        class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                        <Settings2 />
-                                        <span>Outros</span>
-                                        <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                            :class="{ 'rotate-90': open }" />
-                                    </collapsible-trigger>
-                                    <collapsible-content>
-                                        <div class="grid gap-2 pl-6">
-                                            <router-link to="/patrimonio/categorias"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <Contact />
-                                                <span>Tipos</span>
-                                            </router-link>
-                                        </div>
-                                    </collapsible-content>
-                                </collapsible>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                    <collapsible v-if="menuStore.permissions.vendas" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <Tags />
-                            <span>Vendas</span>
-                            <div class="ml-auto flex space-x-1 items-center">
-                                <Badge v-if="true" variant="destructive"> {{ stateDevelopment }} </Badge>
-                                <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                    :class="{ 'rotate-90': open }" />
-                            </div>
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/patrimonio/categorias"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileChartPie />
-                                    <span>Resumo</span>
-                                </router-link>
-                                <router-link to="/vendas/pdv"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <Computer />
-                                    <span>PDV</span>
-                                </router-link>
-                                <router-link to="/vendas/pdv"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <ClipboardCheck />
-                                    <span>Relatórios</span>
-                                </router-link>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                    <collapsible v-if="menuStore.permissions.financeiro" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <Landmark />
-                            <span>Financeiro</span>
-                            <div class="ml-auto flex space-x-1 items-center">
-                                <Badge v-if="true" variant="destructive"> {{ stateDevelopment }} </Badge>
-                                <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                    :class="{ 'rotate-90': open }" />
-                            </div>
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/patrimonio/categorias"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileChartPie />
-                                    <span>Resumo</span>
-                                </router-link>
-                                <router-link to="/financeiro/lancamentos"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <Wallet />
-                                    <span>Lançamentos</span>
-                                </router-link>
-                                <router-link to="/financeiro/dre"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileText />
-                                    <span>DRE</span>
-                                </router-link>
-                                <collapsible v-slot="{ open }" class="grid gap-2">
-                                    <collapsible-trigger
-                                        class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                        <Settings2 />
-                                        <span>Outros</span>
-                                        <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                            :class="{ 'rotate-90': open }" />
-                                    </collapsible-trigger>
-                                    <collapsible-content>
-                                        <div class="grid gap-2 pl-6">
-                                            <router-link to="/financeiro/dre"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <WalletMinimal />
-                                                <span>Contas</span>
-                                            </router-link>
-                                            <router-link to="/financeiro/dre"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <CircleDollarSign />
-                                                <span>Formas pg.</span>
-                                            </router-link>
-                                            <router-link to="/financeiro/dre"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <Group />
-                                                <span>Categorias</span>
-                                            </router-link>
-                                        </div>
-                                    </collapsible-content>
-                                </collapsible>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                    <collapsible v-if="menuStore.permissions.patrimonio" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <Package />
-                            <span>Patrimônio</span>
-                            <div class="ml-auto flex space-x-1 items-center">
-                                <Badge v-if="true" variant="destructive"> {{ stateDevelopment }} </Badge>
-                                <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                    :class="{ 'rotate-90': open }" />
-                            </div>
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/patrimonio/categorias"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileChartPie />
-                                    <span>Resumo</span>
-                                </router-link>
-                                <router-link to="/patrimonio/produtos"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <Archive />
-                                    <span>Produtos</span>
-                                </router-link>
-                                <router-link to="/patrimonio/estoques"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <Boxes />
-                                    <span>Estoques</span>
-                                </router-link>
-                                <collapsible v-slot="{ open }" class="grid gap-2">
-                                    <collapsible-trigger
-                                        class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                        <Settings2 />
-                                        <span>Outros</span>
-                                        <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                            :class="{ 'rotate-90': open }" />
-                                    </collapsible-trigger>
-                                    <collapsible-content>
-                                        <div class="grid gap-2 pl-6">
-                                            <router-link to="/patrimonio/categorias"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <Group />
-                                                <span>Categorias</span>
-                                            </router-link>
-                                            <router-link to="/patrimonio/fornecedores"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <Group />
-                                                <span>Fornecedores</span>
-                                            </router-link>
-                                        </div>
-                                    </collapsible-content>
-                                </collapsible>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                    <collapsible v-if="menuStore.permissions.servicos" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <FileDigit />
-                            <span>Serviços</span>
-                            <div class="ml-auto flex space-x-1 items-center">
-                                <Badge v-if="true" variant="destructive"> {{ stateDevelopment }} </Badge>
-                                <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                    :class="{ 'rotate-90': open }" />
-                            </div>
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/patrimonio/categorias"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileChartPie />
-                                    <span>Resumo</span>
-                                </router-link>
-                                <router-link to="/patrimonio/servicos"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <BookOpenCheck />
-                                    <span>Ordens de Serviço</span>
-                                </router-link>
-                                <router-link to="/patrimonio/servicos"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileCheck />
-                                    <span>Serviços</span>
-                                </router-link>
-                                <collapsible v-slot="{ open }" class="grid gap-2">
-                                    <collapsible-trigger
-                                        class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                        <Settings2 />
-                                        <span>Outros</span>
-                                        <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                            :class="{ 'rotate-90': open }" />
-                                    </collapsible-trigger>
-                                    <collapsible-content>
-                                        <div class="grid gap-2 pl-6">
-                                            <router-link to="/patrimonio/categorias"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <FileBadge2 />
-                                                <span>Garantias</span>
-                                            </router-link>
-                                            <router-link to="/patrimonio/categorias"
-                                                class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                                <FileStack />
-                                                <span>Tipos</span>
-                                            </router-link>
-                                        </div>
-                                    </collapsible-content>
-                                </collapsible>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                    <collapsible v-if="menuStore.permissions.assinantes" v-slot="{ open }" class="grid gap-2">
-                        <collapsible-trigger
-                            class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                            <PenTool />
-                            <span>Assinantes</span>
-                            <div class="ml-auto flex space-x-1 items-center">
-                                <Badge v-if="true" variant="destructive"> {{ stateDevelopment }} </Badge>
-                                <ChevronRight class="ml-auto h-4 w-4 transition-transform"
-                                    :class="{ 'rotate-90': open }" />
-                            </div>
-                        </collapsible-trigger>
-                        <collapsible-content>
-                            <div class="grid gap-2 pl-6">
-                                <router-link to="/patrimonio/categorias"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <FileChartPie />
-                                    <span>Resumo</span>
-                                </router-link>
-                                <router-link to="/patrimonio/produtos"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <Ticket />
-                                    <span>Assinaturas</span>
-                                </router-link>
-                                <router-link to="/patrimonio/produtos"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <BadgeCheck />
-                                    <span>Planos</span>
-                                </router-link>
-                                <router-link to="/patrimonio/servicos"
-                                    class="flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium transition-colors hover:bg-sidebar-hover">
-                                    <CalendarX2 />
-                                    <span>Vencimentos</span>
-                                </router-link>
-                            </div>
-                        </collapsible-content>
-                    </collapsible>
-                </nav>
-            </div>
+            <Sidebar />
             <div class="mt-auto px-3 py-4 bg-sidebar">
                 <Card v-if="!loginStore.isProAccount && loginStore.isAdminUser">
                     <CardHeader class="text-center">
@@ -395,6 +52,7 @@
                 </div>
             </header>
             <main class="p-4 lg:p-6">
+                <GlobalAlert class="mb-2" />
                 <router-view v-if="showComponent" v-slot="{ Component }">
                     <transition name="fade" @before-enter="beforeEnter" @after-leave="afterLeave">
                         <component :is="Component" />
@@ -407,24 +65,20 @@
 
 <script setup lang="ts">
 import { ref, computed } from 'vue';
-import { House, LockKeyhole, Users, List, Wallet, ChevronRight, LogOut, PanelLeftOpen, PanelLeftClose, Landmark, FileText, Package, Boxes, Group, Computer, Tags, BookOpenCheck, BadgeCheck, CalendarX2, FileChartPie, Settings2, PenTool, ClipboardCheck, Ticket, FileDigit, Archive, FileBadge2, FileStack, FileCheck, Layers, User, Contact, WalletMinimal, CircleDollarSign, Crown, Cpu, MessageCircle } from 'lucide-vue-next'
-import { Collapsible, CollapsibleTrigger, CollapsibleContent } from "@/components/ui/collapsible"
+import { LogOut, PanelLeftOpen, PanelLeftClose } from 'lucide-vue-next'
 import { BreadCrumb, ProfileHome, ToogleMode } from '.';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { LoginService } from '@/services/login/loginService';
 import { onMounted } from 'vue';
 import { onUnmounted } from 'vue';
-import { Badge } from '@/components/ui/badge';
 import { useLoginStore } from '@/stores/login/loginStore';
-import { useMenuStore } from '@/stores/menuStore';
-import { Skeleton } from '@/components/ui/skeleton';
+
+import Sidebar from './sidebar.vue';
+import GlobalAlert from './globalAlert.vue';
 
 const widthWindow = ref(window.innerWidth);
-const stateDevelopment = ref('beta');
-
 const loginStore = useLoginStore();
-const menuStore = useMenuStore();
 
 const handleResize = () => {
     widthWindow.value = window.innerWidth;
