+++ vue/src/layouts/shadcn/stateLayout.ts
@@ -1,0 +1,14 @@
+import { defineStore } from "pinia";
+import { ref } from "vue";
+
+export const useLayoutStore = defineStore("stateLayout", () => {
+    const isOnline = ref<boolean>(navigator.onLine);
+    window.addEventListener("online", () => {
+        isOnline.value = true;
+    })
+
+    window.addEventListener("offline", () => {
+        isOnline.value = false;
+    })
+    return { isOnline };
+});
